+++ projects/ids-enterprise-ng/src/lib/datagrid/soho-datagrid.d.ts
@@ -557,6 +557,7 @@ interface SohoStatic {
     Ellipsis: SohoDataGridColumnFormatterFunction;
     Password: SohoDataGridColumnFormatterFunction;
     Readonly: SohoDataGridColumnFormatterFunction;
+    RowNumber: SohoDataGridColumnFormatterFunction;
     Date: SohoDataGridColumnFormatterFunction;
     Time: SohoDataGridColumnFormatterFunction;
     Autocomplete: SohoDataGridColumnFormatterFunction;